% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{wise_diagnose}
\alias{wise_diagnose}
\title{Diagnose an incompatible WISE-Surv class}
\usage{
wise_diagnose(object)
}
\arguments{
\item{object}{A fitted \code{wise_surv_fit} object.}
}
\value{
A list containing the total and group-specific minimum relaxation.
}
\description{
Diagnose an incompatible WISE-Surv class
}
