% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{plot.wise_bounds}
\alias{plot.wise_bounds}
\title{Plot extremizing trajectories}
\usage{
\method{plot}{wise_bounds}(x, which = 1L, ...)
}
\arguments{
\item{x}{A \code{wise_bounds} object.}

\item{which}{Row of the bounds summary to plot.}

\item{...}{Additional arguments passed to \code{plot}.}
}
\description{
Plot extremizing trajectories
}
